net: add efihttp to do HTTP(S) Boot by UEFI HTTP Protocol
This patch implements HTTP Boot by using UEFI HTTP Protocols
defined in the UEFI spec (v2.5+):

 28.6 EFI HTTP Protocols

It would be better to make efihttp as a single module similar to http.mod.
But I cannot fit my implementation to the interfaces (grub_net_app_protocol).
Some operations conflict: there is packets_pulled(), but no read().
So I placed efihttp in the net.mod.

It will be much appreciated to have suggestions of the best way
to get this patch integrated in grub.

Signed-off-by: Ken Lin <ken.lin@hpe.com>
Signed-off-by: Clay Chang <clayc@hpe.com>
Reviewed-by: Keng-Yu Lin <kengyu@hpe.com>